Title: n개의 숫자 중 최대 2개, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240501/n개의 숫자 중 최대 2개/two-max-of-n-num.py
@@ -1,5 +1,5 @@
 n = int(input())
-arr = list(map(int,input().split()))
+arr = list(map(long,input().split()))
 for i in range(0,n):
     for j in range(n - i - 1):        
         if arr[i] < arr[i+j]:
